Added Login Validation
I have altered the GUI so it has the ability to show the users an error
if they give the wrong initial Online Login details. Much of the code
will be able to be reused for the standard Passcode login.

The validation check is done in a method called loginCheck which returns
a boolean value which returns true currently to skip to the next screen