Added Debug Start
Added variables to the SpashActivity class which control the type of
start you want. This enables the initial setup or the login to be
skipped each time the app is run in order to help others when testing
the app.

There are three starts:
- Normal Start which simply acts as if the user is opening the app, it
will take you to the standard passcode login screen

- First Time Start which makes the app act as if it is being setup for
the first time ever

- Skip Start which skips all of the login options and goes straight to
the home tabs

To use these options simply enable the one you want by changing its
boolean value to true and all of the others to false