*Read Description*
Fixed PHPHandler - Which now works but still fragile and still needs
work.

Validation will now start when 4 characters have been entered.

I have also added a loading dialog to the async task which is the
PHPHandler class. Now a loading icon appears with a transparent
background when the process is being performed