Setting Page - Feedback Screen
- button has been implemented, for the time being I have set up so that
it get the value from the screen and send it to the email app. but if we
have time I can change this and make it look like a background activity
where the feedback will be send to an email address when the user clicks
on the button. so no need to show it to the user. but it will take a bit
of time so I will leave it to the end