Updated Styling Info
I have added a styles.xml file which we will begin to use to ensure
consistency across the application and make changes to the application
easier. The file works similar to the string and colours resource files
and a CSS file. I will be giving instructions to the design team when
more of the design is complete to use this.