Fixed Keyboard Bug Add Cash Entry
I've fixed a bug which meant that on phone running certain APIs you were
unable to change move on to the other inputs. I have now implemented an
Input Manager which deals with hiding and showing the keyboard.

As part of this the inputs have been set to auto skip to the next input
when the user has finished in a similar way to pressing tab on a PC.

If people could all take the time to try out the feature that would be
great as the bug only affected certain APIs before and whilst I think it
is now fixed for all the more levels it is tested on the better, let me
know if you've got any feedback