Revised Login Activity
As requested I have changed the login screen to look more professional.
The screen now has four single character inputs which are combined in
the Java file into the passcode which is passed to the server. The input
switching between boxes is all automatic and the keyboard managed by an
input manager.

As well as this I have added comments and optimised the code and XML
files so they fit with the rest of the applciation

NOTE: I have only edited Arty's code basically as I have little
understanding, it appears to work for me but if it could be checked by
someone who understands that would be great