Change Passcode GUI Finished
The change Passcode area of the settings is now functioning and
complete. In sort it takes two Passcodes from the user, the one they
want to swap to twice to confirm and then displays a dialog which asks
for the old password, once that is correctly inputted then the Passcode
calls a method which will update the database.

The backend still needs implementing for this